requirements.txt
bs4
lxml
requests
discord.py
python-dotenv